Wrapper function for toolmarks data set in tmaRks package

This function is specific to thetoolmarks dataset. It unpacks the information stored in the tool ID. This allows us to store the data in a smaller dataset but unpack it for the analysis.

data(
  ...,
  list = character(),
  package = NULL,
  lib.loc = NULL,
  verbose = getOption("verbose"),
  envir = .GlobalEnv,
  overwrite = TRUE
)

Arguments

...

literal character strings or names.

list

a character vector.

package

a character vector giving the package(s) to look in for data sets, or NULL.

By default, all packages in the search path are used, then the data subdirectory (if present) of the current working directory.

lib.loc

a character vector of directory names of R libraries, or NULL. The default value of NULL corresponds to all libraries currently known.

verbose

a logical. If TRUE, additional diagnostics are printed.

envir

the environment where the data should be loaded.

overwrite

logical: should existing objects of the same name in envir be replaced?

Examples

data(toolmarks)
dim(toolmarks)
#> [1] 386712     11